requirements.txt
langchain
langchain-openai
python-dotenv
openai
flask
flask-cors
beautifulsoup4
requests
markdown
